requirements.txt
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
idna==2.9
python-dotenv==0.12.0
requests==2.23.0
soupsieve==2.0
urllib3==1.25.8
